Refactor LeaderboardList component to utilize FILTER_OPTIONS for filtering and improve data fetching logic. Removed unused code and integrated LeaderboardRow for rendering participants.
